Beginnings of java.util.logging adaptation
The beginnings of code to send Redwood output through other
services, e.g., java.util.logging. The version implemented
here will send output through java.util.logging, and has the
framework to send it through other means.

Missing is the ability to do this easily from RedwoodConfiguration,
and perhaps more importantly the ability to do this if
Redwood is already capturing java.util.logging -- wherein the
Level transformation is lossy, tracks may be a bit strange,
and there's a chance for infinite loops.